--HG--
files : cms/models/pagemodel.py
files : example/settings.py

+++ cms/models/pagemodel.py
@@ -294,7 +294,12 @@ class Page(Publisher, Mptt):
             path = self.get_slug(language, fallback)
         else:
             path = self.get_path(language, fallback)
-            if self.parent_id and self.get_cached_ancestors()[0].pk == self.get_home_pk_cache():
+            home_pk = None
+            try:
+                home_pk = self.get_home_pk_cache()
+            except NoHomeFound:
+                pass
+            if self.parent_id and self.get_cached_ancestors()[0].pk == home_pk:
                 path = "/".join(path.split("/")[1:])
             
         return urljoin(reverse('pages-root'), path)
@@ -516,13 +521,21 @@ class Page(Publisher, Mptt):
         if self.parent_id:
             return False
         else:
-            return self.get_home_pk_cache() == self.pk
+            try:
+                return self.get_home_pk_cache() == self.pk
+            except NoHomeFound:
+                pass
+        return False
         
     def is_parent_home(self):
         if not self.parent_id:
             return False
         else:
-            return self.get_home_pk_cache() == self.parent_id
+            try:
+                return self.get_home_pk_cache() == self.parent_id
+            except NoHomeFound:
+                pass
+        return False
         
     def get_home_pk_cache(self):
         if not hasattr(self, "home_pk_cache"):
@@ -586,6 +599,9 @@ class Page(Publisher, Mptt):
         
         Returns: True if page was successfully published.
         """
+        if not settings.CMS_MODERATOR:
+            return
+        
         # clean moderation log
         self.pagemoderatorstate_set.all().delete()
